Refactored
- EnhancedProcessManager for intelligent resource allocation and command execution.
- PerformanceDashboard for real-time performance tracking and historical execution data.
- ProcessPool with auto-scaling capabilities to manage task execution efficiently.
- ResourceMonitor for tracking system resource usage and trends.
- Caching mechanism with AdvancedCache to optimize command result retrieval.
- Refactored hexstrike_server to utilize the new process management architecture.
