+++ core/advanced_cache.py
@@ -1,0 +1,129 @@
+import threading
+import time
+from typing import Any, Optional, Dict
+
+import logging
+logger = logging.getLogger(__name__)
+
+class AdvancedCache:
+    """Advanced caching system with intelligent TTL and LRU eviction"""
+
+    def __init__(self, max_size=1000, default_ttl=3600):
+        self.max_size = max_size
+        self.default_ttl = default_ttl
+        self.cache = {}
+        self.access_times = {}
+        self.ttl_times = {}
+        self.cache_lock = threading.RLock()
+        self.hit_count = 0
+        self.miss_count = 0
+
+        # Start cleanup thread
+        self.cleanup_thread = threading.Thread(target=self._cleanup_expired, daemon=True)
+        self.cleanup_thread.start()
+
+    def get(self, key: str) -> Any:
+        """Get value from cache"""
+        with self.cache_lock:
+            current_time = time.time()
+
+            # Check if key exists and is not expired
+            if key in self.cache and (key not in self.ttl_times or self.ttl_times[key] > current_time):
+                # Update access time for LRU
+                self.access_times[key] = current_time
+                self.hit_count += 1
+                return self.cache[key]
+
+            # Cache miss or expired
+            if key in self.cache:
+                # Remove expired entry
+                self._remove_key(key)
+
+            self.miss_count += 1
+            return None
+
+    def set(self, key: str, value: Any, ttl: Optional[int] = None) -> None:
+        """Set value in cache with optional TTL"""
+        with self.cache_lock:
+            current_time = time.time()
+
+            # Use default TTL if not specified
+            if ttl is None:
+                ttl = self.default_ttl
+
+            # Check if we need to evict entries
+            if len(self.cache) >= self.max_size and key not in self.cache:
+                self._evict_lru()
+
+            # Set the value
+            self.cache[key] = value
+            self.access_times[key] = current_time
+            self.ttl_times[key] = current_time + ttl
+
+    def delete(self, key: str) -> bool:
+        """Delete key from cache"""
+        with self.cache_lock:
+            if key in self.cache:
+                self._remove_key(key)
+                return True
+            return False
+
+    def clear(self) -> None:
+        """Clear all cache entries"""
+        with self.cache_lock:
+            self.cache.clear()
+            self.access_times.clear()
+            self.ttl_times.clear()
+
+    def _remove_key(self, key: str) -> None:
+        """Remove key and associated metadata"""
+        self.cache.pop(key, None)
+        self.access_times.pop(key, None)
+        self.ttl_times.pop(key, None)
+
+    def _evict_lru(self) -> None:
+        """Evict least recently used entry"""
+        if not self.access_times:
+            return
+
+        # Find least recently used key
+        lru_key = min(self.access_times.keys(), key=lambda k: self.access_times[k])
+        self._remove_key(lru_key)
+        logger.debug(f"🗑️ Evicted LRU cache entry: {lru_key}")
+
+    def _cleanup_expired(self) -> None:
+        """Cleanup expired entries periodically"""
+        while True:
+            try:
+                time.sleep(60)  # Cleanup every minute
+                current_time = time.time()
+                expired_keys = []
+
+                with self.cache_lock:
+                    for key, expiry_time in self.ttl_times.items():
+                        if expiry_time <= current_time:
+                            expired_keys.append(key)
+
+                    for key in expired_keys:
+                        self._remove_key(key)
+
+                if expired_keys:
+                    logger.debug(f"🧹 Cleaned up {len(expired_keys)} expired cache entries")
+
+            except Exception as e:
+                logger.error(f"💥 Cache cleanup error: {str(e)}")
+
+    def get_stats(self) -> Dict[str, Any]:
+        """Get cache statistics"""
+        with self.cache_lock:
+            total_requests = self.hit_count + self.miss_count
+            hit_rate = (self.hit_count / total_requests * 100) if total_requests > 0 else 0
+
+            return {
+                "size": len(self.cache),
+                "max_size": self.max_size,
+                "hit_count": self.hit_count,
+                "miss_count": self.miss_count,
+                "hit_rate": hit_rate,
+                "utilization": (len(self.cache) / self.max_size * 100)
+            }
