+++ core/enhanced_process_manager.py
@@ -1,0 +1,229 @@
+import threading
+import subprocess
+import os
+import time
+from typing import Dict, Any, Optional
+from process_pool import ProcessPool
+from advanced_cache import AdvancedCache
+from resource_monitor import ResourceMonitor
+from performance_dashboard import PerformanceDashboard
+
+import logging
+logger = logging.getLogger(__name__)
+
+class EnhancedProcessManager:
+    """Advanced process management with intelligent resource allocation"""
+
+    def __init__(self):
+        self.process_pool = ProcessPool(min_workers=4, max_workers=32)
+        self.cache = AdvancedCache(max_size=2000, default_ttl=1800)  # 30 minutes default TTL
+        self.resource_monitor = ResourceMonitor()
+        self.process_registry = {}
+        self.registry_lock = threading.RLock()
+        self.performance_dashboard = PerformanceDashboard()
+
+        # Process termination and recovery
+        self.termination_handlers = {}
+        self.recovery_strategies = {}
+
+        # Auto-scaling configuration
+        self.auto_scaling_enabled = True
+        self.resource_thresholds = {
+            "cpu_high": 85.0,
+            "memory_high": 90.0,
+            "disk_high": 95.0,
+            "load_high": 0.8
+        }
+
+        # Start background monitoring
+        self.monitor_thread = threading.Thread(target=self._monitor_system, daemon=True)
+        self.monitor_thread.start()
+
+    def execute_command_async(self, command: str, context: Optional[Dict[str, Any]] = None) -> str:
+        """Execute command asynchronously using process pool"""
+        task_id = f"cmd_{int(time.time() * 1000)}_{hash(command) % 10000}"
+
+        # Check cache first
+        cache_key = f"cmd_result_{hash(command)}"
+        cached_result = self.cache.get(cache_key)
+        if cached_result and context and context.get("use_cache", True):
+            logger.info(f"📋 Using cached result for command: {command[:50]}...")
+            return cached_result
+
+        # Submit to process pool
+        self.process_pool.submit_task(
+            task_id,
+            self._execute_command_internal,
+            command,
+            context or {}
+        )
+
+        return task_id
+
+    def _execute_command_internal(self, command: str, context: Dict[str, Any]) -> Dict[str, Any]:
+        """Internal command execution with enhanced monitoring"""
+        start_time = time.time()
+        process = None
+
+        try:
+            # Resource-aware execution
+            resource_usage = self.resource_monitor.get_current_usage()
+
+            # Adjust command based on resource availability
+            if resource_usage["cpu_percent"] > self.resource_thresholds["cpu_high"]:
+                # Add nice priority for CPU-intensive commands
+                if not command.startswith("nice"):
+                    command = f"nice -n 10 {command}"
+
+            # Execute command
+            process = subprocess.Popen(
+                command,
+                shell=True,
+                stdout=subprocess.PIPE,
+                stderr=subprocess.PIPE,
+                text=True,
+                preexec_fn=os.setsid if os.name != 'nt' else None
+            )
+
+            # Register process
+            with self.registry_lock:
+                self.process_registry[process.pid] = {
+                    "command": command,
+                    "process": process,
+                    "start_time": start_time,
+                    "context": context,
+                    "status": "running"
+                }
+
+            # Monitor process execution
+            stdout, stderr = process.communicate()
+            execution_time = time.time() - start_time
+
+            result = {
+                "success": process.returncode == 0,
+                "stdout": stdout,
+                "stderr": stderr,
+                "return_code": process.returncode,
+                "execution_time": execution_time,
+                "pid": process.pid,
+                "resource_usage": self.resource_monitor.get_process_usage(process.pid)
+            }
+
+            # Cache successful results
+            if result["success"] and context.get("cache_result", True):
+                cache_key = f"cmd_result_{hash(command)}"
+                cache_ttl = context.get("cache_ttl", 1800)  # 30 minutes default
+                self.cache.set(cache_key, result, cache_ttl)
+
+            # Update performance metrics
+            self.performance_dashboard.record_execution(command, result)
+
+            return result
+
+        except Exception as e:
+            execution_time = time.time() - start_time
+            error_result = {
+                "success": False,
+                "stdout": "",
+                "stderr": str(e),
+                "return_code": -1,
+                "execution_time": execution_time,
+                "error": str(e)
+            }
+
+            self.performance_dashboard.record_execution(command, error_result)
+            return error_result
+
+        finally:
+            # Cleanup process registry
+            with self.registry_lock:
+                if process is not None and hasattr(process, 'pid') and process.pid in self.process_registry:
+                    del self.process_registry[process.pid]
+
+    def get_task_result(self, task_id: str) -> Dict[str, Any]:
+        """Get result of async task"""
+        return self.process_pool.get_task_result(task_id)
+
+    def terminate_process_gracefully(self, pid: int, timeout: int = 30) -> bool:
+        """Terminate process with graceful degradation"""
+        try:
+            with self.registry_lock:
+                if pid not in self.process_registry:
+                    return False
+
+                process_info = self.process_registry[pid]
+                process = process_info["process"]
+
+                # Try graceful termination first
+                process.terminate()
+
+                # Wait for graceful termination
+                try:
+                    process.wait(timeout=timeout)
+                    process_info["status"] = "terminated_gracefully"
+                    logger.info(f"✅ Process {pid} terminated gracefully")
+                    return True
+                except subprocess.TimeoutExpired:
+                    # Force kill if graceful termination fails
+                    process.kill()
+                    process_info["status"] = "force_killed"
+                    logger.warning(f"⚠️ Process {pid} force killed after timeout")
+                    return True
+
+        except Exception as e:
+            logger.error(f"💥 Error terminating process {pid}: {str(e)}")
+            return False
+
+    def _monitor_system(self):
+        """Monitor system resources and auto-scale"""
+        while True:
+            try:
+                time.sleep(15)  # Monitor every 15 seconds
+
+                # Get current resource usage
+                resource_usage = self.resource_monitor.get_current_usage()
+
+                # Auto-scaling based on resource usage
+                if self.auto_scaling_enabled:
+                    self._auto_scale_based_on_resources(resource_usage)
+
+                # Update performance dashboard
+                self.performance_dashboard.update_system_metrics(resource_usage)
+
+            except Exception as e:
+                logger.error(f"💥 System monitoring error: {str(e)}")
+
+    def _auto_scale_based_on_resources(self, resource_usage: Dict[str, float]):
+        """Auto-scale process pool based on resource usage"""
+        pool_stats = self.process_pool.get_pool_stats()
+        current_workers = pool_stats["active_workers"]
+
+        # Scale down if resources are constrained
+        if (resource_usage["cpu_percent"] > self.resource_thresholds["cpu_high"] or
+            resource_usage["memory_percent"] > self.resource_thresholds["memory_high"]):
+
+            if current_workers > self.process_pool.min_workers:
+                self.process_pool._scale_down(1)
+                logger.info(f"📉 Auto-scaled down due to high resource usage: CPU {resource_usage['cpu_percent']:.1f}%, Memory {resource_usage['memory_percent']:.1f}%")
+
+        # Scale up if resources are available and there's demand
+        elif (resource_usage["cpu_percent"] < 60 and
+              resource_usage["memory_percent"] < 70 and
+              pool_stats["queue_size"] > 2):
+
+            if current_workers < self.process_pool.max_workers:
+                self.process_pool._scale_up(1)
+                logger.info(f"📈 Auto-scaled up due to available resources and demand")
+
+    def get_comprehensive_stats(self) -> Dict[str, Any]:
+        """Get comprehensive system and process statistics"""
+        return {
+            "process_pool": self.process_pool.get_pool_stats(),
+            "cache": self.cache.get_stats(),
+            "resource_usage": self.resource_monitor.get_current_usage(),
+            "active_processes": len(self.process_registry),
+            "performance_dashboard": self.performance_dashboard.get_summary(),
+            "auto_scaling_enabled": self.auto_scaling_enabled,
+            "resource_thresholds": self.resource_thresholds
+        }
+
