+++ core/performance_dashboard.py
@@ -1,0 +1,55 @@
+import threading
+import time
+from typing import Dict, Any
+
+
+class PerformanceDashboard:
+    """Real-time performance monitoring dashboard"""
+
+    def __init__(self):
+        self.execution_history = []
+        self.system_metrics = []
+        self.dashboard_lock = threading.Lock()
+        self.max_history = 1000
+
+    def record_execution(self, command: str, result: Dict[str, Any]):
+        """Record command execution for performance tracking"""
+        with self.dashboard_lock:
+            execution_record = {
+                "command": command[:100],  # Truncate long commands
+                "success": result.get("success", False),
+                "execution_time": result.get("execution_time", 0),
+                "return_code": result.get("return_code", -1),
+                "timestamp": time.time()
+            }
+
+            self.execution_history.append(execution_record)
+            if len(self.execution_history) > self.max_history:
+                self.execution_history.pop(0)
+
+    def update_system_metrics(self, metrics: Dict[str, Any]):
+        """Update system metrics for dashboard"""
+        with self.dashboard_lock:
+            self.system_metrics.append(metrics)
+            if len(self.system_metrics) > self.max_history:
+                self.system_metrics.pop(0)
+
+    def get_summary(self) -> Dict[str, Any]:
+        """Get performance summary"""
+        with self.dashboard_lock:
+            if not self.execution_history:
+                return {"executions": 0}
+
+            recent_executions = self.execution_history[-100:]  # Last 100 executions
+
+            total_executions = len(recent_executions)
+            successful_executions = sum(1 for e in recent_executions if e["success"])
+            avg_execution_time = sum(e["execution_time"] for e in recent_executions) / total_executions
+
+            return {
+                "total_executions": len(self.execution_history),
+                "recent_executions": total_executions,
+                "success_rate": (successful_executions / total_executions * 100) if total_executions > 0 else 0,
+                "avg_execution_time": avg_execution_time,
+                "system_metrics_count": len(self.system_metrics)
+            }
