+++ core/resource_monitor.py
@@ -1,0 +1,87 @@
+import threading
+import time
+import psutil
+from typing import Dict, Any
+
+import logging
+logger = logging.getLogger(__name__)
+
+class ResourceMonitor:
+    """Advanced resource monitoring with historical tracking"""
+
+    def __init__(self, history_size=100):
+        self.history_size = history_size
+        self.usage_history = []
+        self.history_lock = threading.Lock()
+
+    def get_current_usage(self) -> Dict[str, float]:
+        """Get current system resource usage"""
+        try:
+            cpu_percent = psutil.cpu_percent(interval=1)
+            memory = psutil.virtual_memory()
+            disk = psutil.disk_usage('/')
+            network = psutil.net_io_counters()
+
+            usage = {
+                "cpu_percent": cpu_percent,
+                "memory_percent": memory.percent,
+                "memory_available_gb": memory.available / (1024**3),
+                "disk_percent": disk.percent,
+                "disk_free_gb": disk.free / (1024**3),
+                "network_bytes_sent": network.bytes_sent,
+                "network_bytes_recv": network.bytes_recv,
+                "timestamp": time.time()
+            }
+
+            # Add to history
+            with self.history_lock:
+                self.usage_history.append(usage)
+                if len(self.usage_history) > self.history_size:
+                    self.usage_history.pop(0)
+
+            return usage
+
+        except Exception as e:
+            logger.error(f"💥 Error getting resource usage: {str(e)}")
+            return {
+                "cpu_percent": 0,
+                "memory_percent": 0,
+                "memory_available_gb": 0,
+                "disk_percent": 0,
+                "disk_free_gb": 0,
+                "network_bytes_sent": 0,
+                "network_bytes_recv": 0,
+                "timestamp": time.time()
+            }
+
+    def get_process_usage(self, pid: int) -> Dict[str, Any]:
+        """Get resource usage for specific process"""
+        try:
+            process = psutil.Process(pid)
+            return {
+                "cpu_percent": process.cpu_percent(),
+                "memory_percent": process.memory_percent(),
+                "memory_rss_mb": process.memory_info().rss / (1024**2),
+                "num_threads": process.num_threads(),
+                "status": process.status()
+            }
+        except (psutil.NoSuchProcess, psutil.AccessDenied):
+            return {}
+
+    def get_usage_trends(self) -> Dict[str, Any]:
+        """Get resource usage trends"""
+        with self.history_lock:
+            if len(self.usage_history) < 2:
+                return {}
+
+            recent = self.usage_history[-10:]  # Last 10 measurements
+
+            cpu_trend = sum(u["cpu_percent"] for u in recent) / len(recent)
+            memory_trend = sum(u["memory_percent"] for u in recent) / len(recent)
+
+            return {
+                "cpu_avg_10": cpu_trend,
+                "memory_avg_10": memory_trend,
+                "measurements": len(self.usage_history),
+                "trend_period_minutes": len(recent) * 15 / 60  # 15 second intervals
+            }
