+++ hexstrike_server.py
@@ -19,11 +19,9 @@ import traceback
 import threading
 import time
 import base64
-import queue
 from concurrent.futures import ThreadPoolExecutor
 from datetime import datetime, timedelta
 from typing import Dict, Any, Optional
-import venv
 from pathlib import Path
 from flask import Flask, request, jsonify
 import psutil
@@ -1869,686 +1867,10 @@ class ParameterOptimizer:
         return recovery_plan
 
 # ============================================================================
-# ADVANCED PROCESS MANAGEMENT AND MONITORING (v10.0 ENHANCEMENT)
+# ADVANCED PROCESS MANAGEMENT AND MONITORING
 # ============================================================================
 
-class ProcessPool:
-    """Intelligent process pool with auto-scaling capabilities"""
-
-    def __init__(self, min_workers=2, max_workers=20, scale_threshold=0.8):
-        self.min_workers = min_workers
-        self.max_workers = max_workers
-        self.scale_threshold = scale_threshold
-        self.workers = []
-        self.task_queue = queue.Queue()
-        self.results = {}
-        self.pool_lock = threading.Lock()
-        self.active_tasks = {}
-        self.performance_metrics = {
-            "tasks_completed": 0,
-            "tasks_failed": 0,
-            "avg_task_time": 0.0,
-            "cpu_usage": 0.0,
-            "memory_usage": 0.0
-        }
-
-        # Initialize minimum workers
-        self._scale_up(self.min_workers)
-
-        # Start monitoring thread
-        self.monitor_thread = threading.Thread(target=self._monitor_performance, daemon=True)
-        self.monitor_thread.start()
-
-    def submit_task(self, task_id: str, func, *args, **kwargs) -> str:
-        """Submit a task to the process pool"""
-        task = {
-            "id": task_id,
-            "func": func,
-            "args": args,
-            "kwargs": kwargs,
-            "submitted_at": time.time(),
-            "status": "queued"
-        }
-
-        with self.pool_lock:
-            self.active_tasks[task_id] = task
-            self.task_queue.put(task)
-
-        logger.info(f"📋 Task submitted to pool: {task_id}")
-        return task_id
-
-    def get_task_result(self, task_id: str) -> Dict[str, Any]:
-        """Get result of a submitted task"""
-        with self.pool_lock:
-            if task_id in self.results:
-                return self.results[task_id]
-            elif task_id in self.active_tasks:
-                return {"status": self.active_tasks[task_id]["status"], "result": None}
-            else:
-                return {"status": "not_found", "result": None}
-
-    def _worker_thread(self, worker_id: int):
-        """Worker thread that processes tasks"""
-        logger.info(f"🔧 Process pool worker {worker_id} started")
-
-        while True:
-            try:
-                # Get task from queue with timeout
-                task = self.task_queue.get(timeout=30)
-                if task is None:  # Shutdown signal
-                    break
-
-                task_id = task["id"]
-                start_time = time.time()
-
-                # Update task status
-                with self.pool_lock:
-                    if task_id in self.active_tasks:
-                        self.active_tasks[task_id]["status"] = "running"
-                        self.active_tasks[task_id]["worker_id"] = worker_id
-                        self.active_tasks[task_id]["started_at"] = start_time
-
-                try:
-                    # Execute task
-                    result = task["func"](*task["args"], **task["kwargs"])
-
-                    # Store result
-                    execution_time = time.time() - start_time
-                    with self.pool_lock:
-                        self.results[task_id] = {
-                            "status": "completed",
-                            "result": result,
-                            "execution_time": execution_time,
-                            "worker_id": worker_id,
-                            "completed_at": time.time()
-                        }
-
-                        # Update performance metrics
-                        self.performance_metrics["tasks_completed"] += 1
-                        self.performance_metrics["avg_task_time"] = (
-                            (self.performance_metrics["avg_task_time"] * (self.performance_metrics["tasks_completed"] - 1) + execution_time) /
-                            self.performance_metrics["tasks_completed"]
-                        )
-
-                        # Remove from active tasks
-                        if task_id in self.active_tasks:
-                            del self.active_tasks[task_id]
-
-                    logger.info(f"✅ Task completed: {task_id} in {execution_time:.2f}s")
-
-                except Exception as e:
-                    # Handle task failure
-                    with self.pool_lock:
-                        self.results[task_id] = {
-                            "status": "failed",
-                            "error": str(e),
-                            "execution_time": time.time() - start_time,
-                            "worker_id": worker_id,
-                            "failed_at": time.time()
-                        }
-
-                        self.performance_metrics["tasks_failed"] += 1
-
-                        if task_id in self.active_tasks:
-                            del self.active_tasks[task_id]
-
-                    logger.error(f"❌ Task failed: {task_id} - {str(e)}")
-
-                self.task_queue.task_done()
-
-            except queue.Empty:
-                # No tasks available, continue waiting
-                continue
-            except Exception as e:
-                logger.error(f"💥 Worker {worker_id} error: {str(e)}")
-
-    def _monitor_performance(self):
-        """Monitor pool performance and auto-scale"""
-        while True:
-            try:
-                time.sleep(10)  # Monitor every 10 seconds
-
-                with self.pool_lock:
-                    queue_size = self.task_queue.qsize()
-                    active_workers = len([w for w in self.workers if w.is_alive()])
-                    active_tasks_count = len(self.active_tasks)
-
-                # Calculate load metrics
-                if active_workers > 0:
-                    load_ratio = (active_tasks_count + queue_size) / active_workers
-                else:
-                    load_ratio = float('inf')
-
-                # Auto-scaling logic
-                if load_ratio > self.scale_threshold and active_workers < self.max_workers:
-                    # Scale up
-                    new_workers = min(2, self.max_workers - active_workers)
-                    self._scale_up(new_workers)
-                    logger.info(f"📈 Scaled up process pool: +{new_workers} workers (total: {active_workers + new_workers})")
-
-                elif load_ratio < 0.3 and active_workers > self.min_workers:
-                    # Scale down
-                    workers_to_remove = min(1, active_workers - self.min_workers)
-                    self._scale_down(workers_to_remove)
-                    logger.info(f"📉 Scaled down process pool: -{workers_to_remove} workers (total: {active_workers - workers_to_remove})")
-
-                # Update performance metrics
-                try:
-                    cpu_percent = psutil.cpu_percent()
-                    memory_info = psutil.virtual_memory()
-
-                    with self.pool_lock:
-                        self.performance_metrics["cpu_usage"] = cpu_percent
-                        self.performance_metrics["memory_usage"] = memory_info.percent
-
-                except Exception:
-                    pass  # Ignore psutil errors
-
-            except Exception as e:
-                logger.error(f"💥 Pool monitor error: {str(e)}")
-
-    def _scale_up(self, count: int):
-        """Add workers to the pool"""
-        with self.pool_lock:
-            for i in range(count):
-                worker_id = len(self.workers)
-                worker = threading.Thread(target=self._worker_thread, args=(worker_id,), daemon=True)
-                worker.start()
-                self.workers.append(worker)
-
-    def _scale_down(self, count: int):
-        """Remove workers from the pool"""
-        with self.pool_lock:
-            for _ in range(count):
-                if len(self.workers) > self.min_workers:
-                    # Signal worker to shutdown by putting None in queue
-                    self.task_queue.put(None)
-                    # Remove from workers list (worker will exit naturally)
-                    if self.workers:
-                        self.workers.pop()
-
-    def get_pool_stats(self) -> Dict[str, Any]:
-        """Get current pool statistics"""
-        with self.pool_lock:
-            active_workers = len([w for w in self.workers if w.is_alive()])
-            return {
-                "active_workers": active_workers,
-                "queue_size": self.task_queue.qsize(),
-                "active_tasks": len(self.active_tasks),
-                "performance_metrics": self.performance_metrics.copy(),
-                "min_workers": self.min_workers,
-                "max_workers": self.max_workers
-            }
-
-class AdvancedCache:
-    """Advanced caching system with intelligent TTL and LRU eviction"""
-
-    def __init__(self, max_size=1000, default_ttl=3600):
-        self.max_size = max_size
-        self.default_ttl = default_ttl
-        self.cache = {}
-        self.access_times = {}
-        self.ttl_times = {}
-        self.cache_lock = threading.RLock()
-        self.hit_count = 0
-        self.miss_count = 0
-
-        # Start cleanup thread
-        self.cleanup_thread = threading.Thread(target=self._cleanup_expired, daemon=True)
-        self.cleanup_thread.start()
-
-    def get(self, key: str) -> Any:
-        """Get value from cache"""
-        with self.cache_lock:
-            current_time = time.time()
-
-            # Check if key exists and is not expired
-            if key in self.cache and (key not in self.ttl_times or self.ttl_times[key] > current_time):
-                # Update access time for LRU
-                self.access_times[key] = current_time
-                self.hit_count += 1
-                return self.cache[key]
-
-            # Cache miss or expired
-            if key in self.cache:
-                # Remove expired entry
-                self._remove_key(key)
-
-            self.miss_count += 1
-            return None
-
-    def set(self, key: str, value: Any, ttl: Optional[int] = None) -> None:
-        """Set value in cache with optional TTL"""
-        with self.cache_lock:
-            current_time = time.time()
-
-            # Use default TTL if not specified
-            if ttl is None:
-                ttl = self.default_ttl
-
-            # Check if we need to evict entries
-            if len(self.cache) >= self.max_size and key not in self.cache:
-                self._evict_lru()
-
-            # Set the value
-            self.cache[key] = value
-            self.access_times[key] = current_time
-            self.ttl_times[key] = current_time + ttl
-
-    def delete(self, key: str) -> bool:
-        """Delete key from cache"""
-        with self.cache_lock:
-            if key in self.cache:
-                self._remove_key(key)
-                return True
-            return False
-
-    def clear(self) -> None:
-        """Clear all cache entries"""
-        with self.cache_lock:
-            self.cache.clear()
-            self.access_times.clear()
-            self.ttl_times.clear()
-
-    def _remove_key(self, key: str) -> None:
-        """Remove key and associated metadata"""
-        self.cache.pop(key, None)
-        self.access_times.pop(key, None)
-        self.ttl_times.pop(key, None)
-
-    def _evict_lru(self) -> None:
-        """Evict least recently used entry"""
-        if not self.access_times:
-            return
-
-        # Find least recently used key
-        lru_key = min(self.access_times.keys(), key=lambda k: self.access_times[k])
-        self._remove_key(lru_key)
-        logger.debug(f"🗑️ Evicted LRU cache entry: {lru_key}")
-
-    def _cleanup_expired(self) -> None:
-        """Cleanup expired entries periodically"""
-        while True:
-            try:
-                time.sleep(60)  # Cleanup every minute
-                current_time = time.time()
-                expired_keys = []
-
-                with self.cache_lock:
-                    for key, expiry_time in self.ttl_times.items():
-                        if expiry_time <= current_time:
-                            expired_keys.append(key)
-
-                    for key in expired_keys:
-                        self._remove_key(key)
-
-                if expired_keys:
-                    logger.debug(f"🧹 Cleaned up {len(expired_keys)} expired cache entries")
-
-            except Exception as e:
-                logger.error(f"💥 Cache cleanup error: {str(e)}")
-
-    def get_stats(self) -> Dict[str, Any]:
-        """Get cache statistics"""
-        with self.cache_lock:
-            total_requests = self.hit_count + self.miss_count
-            hit_rate = (self.hit_count / total_requests * 100) if total_requests > 0 else 0
-
-            return {
-                "size": len(self.cache),
-                "max_size": self.max_size,
-                "hit_count": self.hit_count,
-                "miss_count": self.miss_count,
-                "hit_rate": hit_rate,
-                "utilization": (len(self.cache) / self.max_size * 100)
-            }
-
-class EnhancedProcessManager:
-    """Advanced process management with intelligent resource allocation"""
-
-    def __init__(self):
-        self.process_pool = ProcessPool(min_workers=4, max_workers=32)
-        self.cache = AdvancedCache(max_size=2000, default_ttl=1800)  # 30 minutes default TTL
-        self.resource_monitor = ResourceMonitor()
-        self.process_registry = {}
-        self.registry_lock = threading.RLock()
-        self.performance_dashboard = PerformanceDashboard()
-
-        # Process termination and recovery
-        self.termination_handlers = {}
-        self.recovery_strategies = {}
-
-        # Auto-scaling configuration
-        self.auto_scaling_enabled = True
-        self.resource_thresholds = {
-            "cpu_high": 85.0,
-            "memory_high": 90.0,
-            "disk_high": 95.0,
-            "load_high": 0.8
-        }
-
-        # Start background monitoring
-        self.monitor_thread = threading.Thread(target=self._monitor_system, daemon=True)
-        self.monitor_thread.start()
-
-    def execute_command_async(self, command: str, context: Optional[Dict[str, Any]] = None) -> str:
-        """Execute command asynchronously using process pool"""
-        task_id = f"cmd_{int(time.time() * 1000)}_{hash(command) % 10000}"
-
-        # Check cache first
-        cache_key = f"cmd_result_{hash(command)}"
-        cached_result = self.cache.get(cache_key)
-        if cached_result and context and context.get("use_cache", True):
-            logger.info(f"📋 Using cached result for command: {command[:50]}...")
-            return cached_result
-
-        # Submit to process pool
-        self.process_pool.submit_task(
-            task_id,
-            self._execute_command_internal,
-            command,
-            context or {}
-        )
-
-        return task_id
-
-    def _execute_command_internal(self, command: str, context: Dict[str, Any]) -> Dict[str, Any]:
-        """Internal command execution with enhanced monitoring"""
-        start_time = time.time()
-        process = None
-
-        try:
-            # Resource-aware execution
-            resource_usage = self.resource_monitor.get_current_usage()
-
-            # Adjust command based on resource availability
-            if resource_usage["cpu_percent"] > self.resource_thresholds["cpu_high"]:
-                # Add nice priority for CPU-intensive commands
-                if not command.startswith("nice"):
-                    command = f"nice -n 10 {command}"
-
-            # Execute command
-            process = subprocess.Popen(
-                command,
-                shell=True,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-                text=True,
-                preexec_fn=os.setsid if os.name != 'nt' else None
-            )
-
-            # Register process
-            with self.registry_lock:
-                self.process_registry[process.pid] = {
-                    "command": command,
-                    "process": process,
-                    "start_time": start_time,
-                    "context": context,
-                    "status": "running"
-                }
-
-            # Monitor process execution
-            stdout, stderr = process.communicate()
-            execution_time = time.time() - start_time
-
-            result = {
-                "success": process.returncode == 0,
-                "stdout": stdout,
-                "stderr": stderr,
-                "return_code": process.returncode,
-                "execution_time": execution_time,
-                "pid": process.pid,
-                "resource_usage": self.resource_monitor.get_process_usage(process.pid)
-            }
-
-            # Cache successful results
-            if result["success"] and context.get("cache_result", True):
-                cache_key = f"cmd_result_{hash(command)}"
-                cache_ttl = context.get("cache_ttl", 1800)  # 30 minutes default
-                self.cache.set(cache_key, result, cache_ttl)
-
-            # Update performance metrics
-            self.performance_dashboard.record_execution(command, result)
-
-            return result
-
-        except Exception as e:
-            execution_time = time.time() - start_time
-            error_result = {
-                "success": False,
-                "stdout": "",
-                "stderr": str(e),
-                "return_code": -1,
-                "execution_time": execution_time,
-                "error": str(e)
-            }
-
-            self.performance_dashboard.record_execution(command, error_result)
-            return error_result
-
-        finally:
-            # Cleanup process registry
-            with self.registry_lock:
-                if process is not None and hasattr(process, 'pid') and process.pid in self.process_registry:
-                    del self.process_registry[process.pid]
-
-    def get_task_result(self, task_id: str) -> Dict[str, Any]:
-        """Get result of async task"""
-        return self.process_pool.get_task_result(task_id)
-
-    def terminate_process_gracefully(self, pid: int, timeout: int = 30) -> bool:
-        """Terminate process with graceful degradation"""
-        try:
-            with self.registry_lock:
-                if pid not in self.process_registry:
-                    return False
-
-                process_info = self.process_registry[pid]
-                process = process_info["process"]
-
-                # Try graceful termination first
-                process.terminate()
-
-                # Wait for graceful termination
-                try:
-                    process.wait(timeout=timeout)
-                    process_info["status"] = "terminated_gracefully"
-                    logger.info(f"✅ Process {pid} terminated gracefully")
-                    return True
-                except subprocess.TimeoutExpired:
-                    # Force kill if graceful termination fails
-                    process.kill()
-                    process_info["status"] = "force_killed"
-                    logger.warning(f"⚠️ Process {pid} force killed after timeout")
-                    return True
-
-        except Exception as e:
-            logger.error(f"💥 Error terminating process {pid}: {str(e)}")
-            return False
-
-    def _monitor_system(self):
-        """Monitor system resources and auto-scale"""
-        while True:
-            try:
-                time.sleep(15)  # Monitor every 15 seconds
-
-                # Get current resource usage
-                resource_usage = self.resource_monitor.get_current_usage()
-
-                # Auto-scaling based on resource usage
-                if self.auto_scaling_enabled:
-                    self._auto_scale_based_on_resources(resource_usage)
-
-                # Update performance dashboard
-                self.performance_dashboard.update_system_metrics(resource_usage)
-
-            except Exception as e:
-                logger.error(f"💥 System monitoring error: {str(e)}")
-
-    def _auto_scale_based_on_resources(self, resource_usage: Dict[str, float]):
-        """Auto-scale process pool based on resource usage"""
-        pool_stats = self.process_pool.get_pool_stats()
-        current_workers = pool_stats["active_workers"]
-
-        # Scale down if resources are constrained
-        if (resource_usage["cpu_percent"] > self.resource_thresholds["cpu_high"] or
-            resource_usage["memory_percent"] > self.resource_thresholds["memory_high"]):
-
-            if current_workers > self.process_pool.min_workers:
-                self.process_pool._scale_down(1)
-                logger.info(f"📉 Auto-scaled down due to high resource usage: CPU {resource_usage['cpu_percent']:.1f}%, Memory {resource_usage['memory_percent']:.1f}%")
-
-        # Scale up if resources are available and there's demand
-        elif (resource_usage["cpu_percent"] < 60 and
-              resource_usage["memory_percent"] < 70 and
-              pool_stats["queue_size"] > 2):
-
-            if current_workers < self.process_pool.max_workers:
-                self.process_pool._scale_up(1)
-                logger.info(f"📈 Auto-scaled up due to available resources and demand")
-
-    def get_comprehensive_stats(self) -> Dict[str, Any]:
-        """Get comprehensive system and process statistics"""
-        return {
-            "process_pool": self.process_pool.get_pool_stats(),
-            "cache": self.cache.get_stats(),
-            "resource_usage": self.resource_monitor.get_current_usage(),
-            "active_processes": len(self.process_registry),
-            "performance_dashboard": self.performance_dashboard.get_summary(),
-            "auto_scaling_enabled": self.auto_scaling_enabled,
-            "resource_thresholds": self.resource_thresholds
-        }
-
-class ResourceMonitor:
-    """Advanced resource monitoring with historical tracking"""
-
-    def __init__(self, history_size=100):
-        self.history_size = history_size
-        self.usage_history = []
-        self.history_lock = threading.Lock()
-
-    def get_current_usage(self) -> Dict[str, float]:
-        """Get current system resource usage"""
-        try:
-            cpu_percent = psutil.cpu_percent(interval=1)
-            memory = psutil.virtual_memory()
-            disk = psutil.disk_usage('/')
-            network = psutil.net_io_counters()
-
-            usage = {
-                "cpu_percent": cpu_percent,
-                "memory_percent": memory.percent,
-                "memory_available_gb": memory.available / (1024**3),
-                "disk_percent": disk.percent,
-                "disk_free_gb": disk.free / (1024**3),
-                "network_bytes_sent": network.bytes_sent,
-                "network_bytes_recv": network.bytes_recv,
-                "timestamp": time.time()
-            }
-
-            # Add to history
-            with self.history_lock:
-                self.usage_history.append(usage)
-                if len(self.usage_history) > self.history_size:
-                    self.usage_history.pop(0)
-
-            return usage
-
-        except Exception as e:
-            logger.error(f"💥 Error getting resource usage: {str(e)}")
-            return {
-                "cpu_percent": 0,
-                "memory_percent": 0,
-                "memory_available_gb": 0,
-                "disk_percent": 0,
-                "disk_free_gb": 0,
-                "network_bytes_sent": 0,
-                "network_bytes_recv": 0,
-                "timestamp": time.time()
-            }
-
-    def get_process_usage(self, pid: int) -> Dict[str, Any]:
-        """Get resource usage for specific process"""
-        try:
-            process = psutil.Process(pid)
-            return {
-                "cpu_percent": process.cpu_percent(),
-                "memory_percent": process.memory_percent(),
-                "memory_rss_mb": process.memory_info().rss / (1024**2),
-                "num_threads": process.num_threads(),
-                "status": process.status()
-            }
-        except (psutil.NoSuchProcess, psutil.AccessDenied):
-            return {}
-
-    def get_usage_trends(self) -> Dict[str, Any]:
-        """Get resource usage trends"""
-        with self.history_lock:
-            if len(self.usage_history) < 2:
-                return {}
-
-            recent = self.usage_history[-10:]  # Last 10 measurements
-
-            cpu_trend = sum(u["cpu_percent"] for u in recent) / len(recent)
-            memory_trend = sum(u["memory_percent"] for u in recent) / len(recent)
-
-            return {
-                "cpu_avg_10": cpu_trend,
-                "memory_avg_10": memory_trend,
-                "measurements": len(self.usage_history),
-                "trend_period_minutes": len(recent) * 15 / 60  # 15 second intervals
-            }
-
-class PerformanceDashboard:
-    """Real-time performance monitoring dashboard"""
-
-    def __init__(self):
-        self.execution_history = []
-        self.system_metrics = []
-        self.dashboard_lock = threading.Lock()
-        self.max_history = 1000
-
-    def record_execution(self, command: str, result: Dict[str, Any]):
-        """Record command execution for performance tracking"""
-        with self.dashboard_lock:
-            execution_record = {
-                "command": command[:100],  # Truncate long commands
-                "success": result.get("success", False),
-                "execution_time": result.get("execution_time", 0),
-                "return_code": result.get("return_code", -1),
-                "timestamp": time.time()
-            }
-
-            self.execution_history.append(execution_record)
-            if len(self.execution_history) > self.max_history:
-                self.execution_history.pop(0)
-
-    def update_system_metrics(self, metrics: Dict[str, Any]):
-        """Update system metrics for dashboard"""
-        with self.dashboard_lock:
-            self.system_metrics.append(metrics)
-            if len(self.system_metrics) > self.max_history:
-                self.system_metrics.pop(0)
-
-    def get_summary(self) -> Dict[str, Any]:
-        """Get performance summary"""
-        with self.dashboard_lock:
-            if not self.execution_history:
-                return {"executions": 0}
-
-            recent_executions = self.execution_history[-100:]  # Last 100 executions
-
-            total_executions = len(recent_executions)
-            successful_executions = sum(1 for e in recent_executions if e["success"])
-            avg_execution_time = sum(e["execution_time"] for e in recent_executions) / total_executions
-
-            return {
-                "total_executions": len(self.execution_history),
-                "recent_executions": total_executions,
-                "success_rate": (successful_executions / total_executions * 100) if total_executions > 0 else 0,
-                "avg_execution_time": avg_execution_time,
-                "system_metrics_count": len(self.system_metrics)
-            }
+from core.enhanced_process_manager import EnhancedProcessManager
 
 # Global instances
 tech_detector = TechnologyDetector()
@@ -2570,7 +1892,7 @@ ctf_automator = CTFChallengeAutomator()
 ctf_coordinator = CTFTeamCoordinator()
 
 # ============================================================================
-# PROCESS MANAGEMENT FOR COMMAND TERMINATION (v5.0 ENHANCEMENT)
+# PROCESS MANAGEMENT FOR COMMAND TERMINATION
 # ============================================================================
 
 # Process management for command termination
